pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "avro>=1.12.0",
    "boto3>=1.35.99",
    "botocore>=1.35.0,<1.36",
    "click>=8.1.8",
    "jupysql>=0.11.1",
    "jupysql-plugin>=0.4.5",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.6",
    "msgspec>=0.19.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "polars>=1.26.0",
    "psycopg>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pyalex>=0.18",
    "pyarrow>=19.0.1",
    "pydantic>=2.11.2",
    "pymongo>=4.11.3",
    "pymongoarrow>=1.7.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "sanic>=25.3.0",
    "sqlalchemy>=2.0.40",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
]

[tool.jupysql.SqlMagic]
feedback = true
autopandas = false
